Adds symlink to FileSystem and FakeFileSystem
This is a large change since it means adding a new Node type and all the
corresponding exhaustive matches. I copied the behavior from macOS 11:
A symlink is treated as a file if it ultimately points to a file, and a
Symlynk is treated as a directory if it points to a directory. If
recursing gets you to a broken symlink then it is treated as neither.